QUR-13 Рефакторинг: вынес логику в Stat+ViewModel, добавил методы для изменения даты, добавил анимации и запрет перехода на будущие недели
