+++ .swiftlint.yml
@@ -1,0 +1,4 @@
+disabled_rules:
+ - trailing_whitespace
+identifier_name:
+  min_length: 2
